rename ParkingMapFilters to ParkingMapFilter
